refactor: move keymaps to separate file\n\n- Create lua/keymaps.lua for all keybindings\n- Simplify init.lua to only include essential setup\n- Improve organization and maintainability
